+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat: Measure app start time (#1111)
 - feat: Auto UI Performance Instrumentation (#1105, #1150, #1136, #1139, #1042)
 
 ## 7.1.3